Solved the @Autowired problem
It turns out that what we were trying to solve with a BeanFactoryPostProcessor should be done with an InitializingBean + ApplicationContextAware. A BFPP is called before wiring, and by forcing to create the beans, we circumvented normal wiring. With the new interfaces implemented, we now run in to all sorts of troubles occurring during wiring, which is expected in such a rewrite.