Add suppressing of duplicate messages in groups
If groups are connected in a cycle, messages are forwarded among the
groups, leading to them receiving messages multiple times.

With this commit, a mechanism is added that suppresses duplicates in
groups. A duplicate message is detected by looking at its 'innermost'
message - the message that was originally created by one group member.

In all duplicates, that message has the same URI. In order to allow
for a quick check, we added a member in MessageEventPlaceholder and
a column in the database, holding that URI.

The MessageEventRepository now has a method for retrieving all earlier
messages that have the same innermost message uri, which is used in the
group facet implementation.