Fix a Bugs in owner app message display
* The receiver property for an incoming message was taken from the local message
copy, not from the attached original, which had to fail as the local copy only contains the time stamp

* Now we react to the success messages for our outgoing messages by updating the timestamp of that
the message that was successful in the state. The new timestamp is the receivedTimestamp of the
response message on the server. Thus, all timestamps are server timestamps (on our WoN node) and
should allow for reliable sorting