Set up environment for ProposalToCancel Test, now change query
The query that was was used, and that passed, was the query from
proposals. This should instead be a query for proposeToCancel. The
oneOpenProposal test should not pass, but it was done so that a copy of
the ProposalFunction as ProposalToCancel could be made and tested as a
baseline to work from.