Show message RDF on hover, allow sending one triple
* Message's RDF contents are shown in tooltips when the show RDF mode is
[SHOW]. Note: this only works with messages that were received in the
websocket. Upon Reload, for example, all messages have empty tooltips.

* The special message format
```
::msg:: [predicate] [object]
```
Allows for sending one  triple. The placeholder ::msg:: is replaced
by the message URI, predicate and object are taken as-is and put into
the message. The object must be enclosed in < > if it is to be
interpreted as an URI. Prefixes in predicates
work if they are generally used in the message (e.g. won, msg, rdf, etc.
are supported).