Improve highlevel protocol queries
Many queries were using unconstrained `graph ?g{?s ?p ?o}` clauses
the results of which were not used anywhere. Removing them led to a
considerable speedup.

One test had to be changed in order to still pass after removing that
clause because it had tested for a missing content graph in a message.
However, a missing content graph is supposed to be caught at lower
levels of the protocol stack, so we do not need to check for that here.
Consequently, the complete message was removed in the test case.