Change agreement URI
Agreement URI used to be the URI of the accept message, now it is the
URI of the propose message.

Reason: Using the accept message URI enables the acceptor to bundle
multiple proposals into one agreement; this, however, makes it
impossible to cancel them individually - something the proposer does not
intend but cannot prevent.

This change fixes the problem.