Fixes #1682
The code for retract/reject was prepared to remove any pending
cancellations, but the code never executed because we only checked if
the retracted/rejected message was a propose, not a proposeToCancel.
Adding that check fixes the problem.

While doing that, hashCode() and equals() methods were added to the
ConversationMessage class, using the message URI as the identifying
member.