Change whatsnew query
The problem is that the query we use plus the transformations we apply
cause the call
Query compiledQuery = OpAsQuery.asQuery(q);
in the SparqlMatcherActor.executeQuery() method to swallow the binding
for ?score, although it is clearly present in the Op q. This is an
attempt to bypass that behaviour (bug?).