Fix bug related to bind in whatsaround query.
BIND ends the group graph pattern it comes after. At the place it used
to be, it led to our query transformation to produce a query in which
?score is always unbound in the projection. We could have moved it to
the end of the group graph pattern (before the filters), but putting it
in the projection is even better as we don't risk moving it around so
much in the future.