Adds another button with a different click action for the back button within the post-messages header:
Due to a routing mishap/confusion the desktop view did not close a connection when pressing back but rather returned to the previous page -> and since that page is not necessarily the overview without an open connection (e.g. if you access multiple connections without going back in between), the back button actually does what is intended and opens the previous connection again, which is rather confusing in the desktop view, however this can't happen in the mobile view as the list will be hidden if a connection item has been selected. the router__back in the mobile view is used so we can jump to the correct scroll position (so we do not lose the position the user was before opening the connection)