Add duplicate filter to batchingConsumer
The batching consumer can now be passed a deduplicationKey, which it
puts into a bloom filter. Subsequent calls with the same
deduplicationKey will be ignored, there is a 1% chance that a call with
a different deduplicationKey will be ignored.

The WonWebsocketHandler now uses [atomUri]+[targetAtomUri] as a
deduplication key, this way the notification email counts only one hint
even if multiple matchers send hints.