Change activemq/jms connection config
* acknowledgementModeName=CLIENT_ACKNOWLEDGE
not sure if this is actually required. It makes the activemq consumer
acknowledge each message. Done because the number of messages waiting
for acknowledgement were constantly rising on the server side

* useAsyncSend=true
This (maybe only in connection with the former) does the trick - no more
messages waiting to be acknowledged. Of course, it introduces possible
message loss, but asyncSend is, for most cases, much closer to what we
want.