Change Message URIs and signing
# Message URIs are now content-based

When composing a message, one MUST use the reserved URI `wm:/SELF` as
the message URI. The message is signed in this form. After any signing
is done, a hash is calculated from the whole message dataset. The
algorithm for graph normalization etc. is the same that is used during
message signing. The resulting hash is encoded as a Multihash in Base58
encoding.

# Easy to remember graph names

A message's signature graph has the URI `[msgUri]#signature`, its
envelope graph has the URI `[msgUri]#envelope`. Every message has
exactly one envelope graph, and every message has exactly one signature
graph (except for hints, which have none).

# Fewer signatures

With this commit, it becomes possible to make signatures for two or more
graphs, and this feature is used for most types of messages: they only
have one signature that signs all the graphs in the message. Exception:
CREATE_ATOM and REPLACE messages: their content needs to be verifiable
independently of the message, therefore the content graphs are signed
separately.

# Using Multihash
All hashes are now sha-256 hashes, encoded using multihash, in a Base58
format. Specifically, this pertains to the hash that is signed as part
of the signature process, the key fingerprint, and the message URI.