Remove test for owned atoms
Before sending a push notification, we used to check if the sending atom and the recipient atom belong to the same user. This check does not seem to work for group chats (in which messages are forwarded by the group atom). We remove this check as the overall experience is better without it.