Differentiated error handling while rematching (#3390)
# Current behavior
The matcher-service removes an atom from the rematching index only if it gets a 410 response. Attempts that fail with other errors are not registered as rematch attempts and are therefore checked for rematching at the next rematch tick (every 60s). This leads to constantly rematching atoms with 401 or 403 responses, flooding the node with requests.

# New behavior
The matcher-service removes an atom from the rematching index if it gets a 410 response and registers a rematch attempt for any other error response (4xx or 5xx), thus the usual back-off strategy is followed. This allows for dealing with temporary unavailability or changing access rights for the matcher.